+++ include/player.h
@@ -1,11 +1,22 @@
 #pragma once
 #include "entity.h"
 #include "texture.h"
+#include "time.h"
+
+struct Direction {
+ public:
+  bool up;
+  bool down;
+  bool left;
+  bool right;
+};
 
 class Player : public Entity {
  public:
   Player(Vector2 position, GameTexture texture) : Entity(position, texture) {}
   float MovementSpeed = 5.0f;
+
+  void MovePlayer(Time& Time, Player& Player);
 
  private:
 };
