+++ src/game.cpp
@@ -2,32 +2,7 @@
 
 #include "../include/player.h"
 
-struct Direction {
- public:
-  bool up;
-  bool down;
-  bool left;
-  bool right;
-};
-
-Direction dir;
-static void MovePlayer(Time& Time, Player& Player) {
-  if (!dir.up && !dir.down && !dir.left && !dir.right) {
-    return;
-  }
-
-  Vector2 Position = Player.GetPosition();
-
-  if (dir.up)
-    Position.y -= Player.MovementSpeed * Time.deltaTime * 100;
-  if (dir.down)
-    Position.y += Player.MovementSpeed * Time.deltaTime * 100;
-  if (dir.left)
-    Position.x -= Player.MovementSpeed * Time.deltaTime * 100;
-  if (dir.right)
-    Position.x += Player.MovementSpeed * Time.deltaTime * 100;
-  Player.SetPosition(Position);
-}
+extern Direction dir;
 
 void Game::HandleSDLEvents(Time& Time, Player& Player) {
   while (SDL_PollEvent(&currentevent)) {
